requirements.txt
instabot==0.60.0
Pillow==6.1.0
python-dotenv==0.10.3
requests==2.22.0
